+++ CHANGELOG.md
@@ -49,6 +49,7 @@
 - Default Buffer and ND algebras are now Ops and lack neutral elements (0, 1) as well as algebra-level shapes.
 - Tensor algebra takes read-only structures as input and inherits AlgebraND
 - `UnivariateDistribution` renamed to `Distribution1D`
+- Rework of histograms.
 
 ### Deprecated
 - Specialized `DoubleBufferAlgebra`